Add reaction to Trace Activation IE
The reaction is Trace Failure Indication with misc cause: "unspecified"
It is sent in response to the Tace Activation IE within the Initial
Context Setup Request message and Trace Start message.

It is for furhter study if there is a better cause to explain that
srsENB does not support trace activation. Also is it really required to
react on the Trace Activation IE withn the Initial Context Setup
Request message.